% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instances.R
\name{load_instances}
\alias{load_instances}
\title{Load a batch of instances from JSON files}
\usage{
load_instances(dir, pattern = "*.json", n_instances = NULL)
}
\arguments{
\item{dir}{Directory containing instance JSON files.}

\item{pattern}{Glob pattern for file names (default all \verb{*.json}).}

\item{n_instances}{Maximum number of instances to load (NULL = all).}
}
\value{
Named list of instance lists, one per file.
}
\description{
Load a batch of instances from JSON files
}
