% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{plot_sensitivity}
\alias{plot_sensitivity}
\title{Plot sensitivity density curves}
\usage{
plot_sensitivity(optx, title = "Sensitivity", scale_x = 1000)
}
\arguments{
\item{optx}{Matrix from \code{\link[=sensitivity_inflow]{sensitivity_inflow()}} or \code{\link[=sensitivity_spot]{sensitivity_spot()}}.}

\item{title}{Plot title.}

\item{scale_x}{Divisor for x-axis labels (default 1000).}
}
\description{
Plot sensitivity density curves
}
