% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logs.R
\name{summarise_logs}
\alias{summarise_logs}
\title{Summarise log results}
\usage{
summarise_logs(log_results)
}
\arguments{
\item{log_results}{Matrix from \code{parse_logs()}.}
}
\value{
Named list with mean/CI for time, bound bias, and CI ratio.
}
\description{
Summarise log results
}
